# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MediMarket Multi-Vendor Medical Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersListResponse"
  /users/{id}:
    get:
      operationId: getUserById
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/status:
    patch:
      operationId: updateUserStatus
      tags: [users]
      summary: Activate or deactivate user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserStatusUpdate"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /stores:
    get:
      operationId: listStores
      tags: [stores]
      summary: List medical stores
      parameters:
        - name: verified
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of stores
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoresListResponse"
    post:
      operationId: createStore
      tags: [stores]
      summary: Create a medical store (store owner only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreInput"
      responses:
        "201":
          description: Store created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
  /stores/{id}:
    get:
      operationId: getStoreById
      tags: [stores]
      summary: Get store by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Store found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
    patch:
      operationId: updateStore
      tags: [stores]
      summary: Update store details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [customer, store_owner, delivery_partner, admin]
        avatar:
          type: ["string", "null"]
        isActive:
          type: boolean
        isVerified:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - isActive
        - isVerified
        - createdAt
...
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        avatar:
          type: string
    UserStatusUpdate:
      type: object
      properties:
        isActive:
          type: boolean
      required:
        - isActive
    UsersListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - users
        - total
        - page
        - limit
    Store:
      type: object
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        licenseNumber:
          type: ["string", "null"]
        isVerified:
          type: boolean
        isActive:
          type: boolean
        rating:
          type: ["number", "null"]
        totalOrders:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - ownerId
        - name
        - isVerified
        - isActive
        - totalOrders
        - createdAt
    StoreInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        description:
          type: string
        address:
          type: string
        city:
          type: string
        licenseNumber:
          type: string
      required:
        - name
    StoreUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        address:
          type: string
        city:
          type: string
        licenseNumber:
          type: string
    StoresListResponse:
      type: object
      properties:
        stores:
          type: array
          items:
            $ref: "#/components/schemas/Store"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - stores
        - total
        - page
        - limit
...