# lib/api-zod/src/generated/types/usersListResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MediMarket Multi-Vendor Medical Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface UsersListResponse {
  users: User[];
  total: number;
  page: number;
  limit: number;
}